requirements.txt
selenium
webdriver-manager
pytest
allure-pytest
allure-python-commons
python-dotenv
faker
pytest-xdist
pytest-rerunfailures